"source" resources require isopack-2 format
(Specifically, the isopack-2-unibuild unibuild format which only lives
in isopack-2 isopacks.)

When saving an isopack, we always save as isopack-2, and if the only
"source"s are CSS and JS we also try to make an isopack-1 version if
we're publishing the package (or saving it to our tropohouse).  If we're
just saving it to an app-specific IsopackCache, we don't bother to
create the old version (if an old version of the tool sees this app's
IsopackCache it will have a non-matching BUILT_BY and rebuild the
isopack).

This commit only implements the CSS side of things; the JS side will be
harder (we'll be putting multiple JS sources into the new isopack but
one prelink resource into the old one).